# docs/接口契约/router.openapi.yaml
openapi: 3.1.0
info:
  title: Router API
  version: 1.0.0
  description: |
    Router HTTP API specification.

    This file is the source of truth for API paths, authentication schemes,
    high-level request bodies, and response envelopes. Business semantics are
    documented in the themed documents under docs/.
servers:
  - url: http://localhost:3000
    description: Local Router server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/admin/channels:
    get:
      tags: [Admin Channels]
      summary: List channels
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Keyword"
      responses:
        "200": { $ref: "#/components/responses/PaginatedAPIResponse" }
...
  /api/v1/admin/channel/{id}:
    get:
      tags: [Admin Channels]
      summary: Get channel detail
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200": { $ref: "#/components/responses/APIResponse" }
    delete:
      tags: [Admin Channels]
      summary: Delete channel
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200": { $ref: "#/components/responses/APIResponse" }
  /api/v1/admin/channel:
    post:
      tags: [Admin Channels]
      summary: Create channel
      requestBody: { $ref: "#/components/requestBodies/JSONBody" }
      responses:
        "200": { $ref: "#/components/responses/APIResponse" }
    put:
      tags: [Admin Channels]
      summary: Update channel
      requestBody: { $ref: "#/components/requestBodies/JSONBody" }
      responses:
        "200": { $ref: "#/components/responses/APIResponse" }
...
  /api/v1/admin/channel/{id}/billing:
    get:
      tags: [Admin Channels]
      summary: Get channel billing summary
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200": { $ref: "#/components/responses/APIResponse" }
...
  /api/v1/admin/channel/{id}/billing/reconciliation:
    get:
      tags: [Admin Channels]
      summary: Reconcile provider usage with Router request logs
      description: Read-only matching by channel, model, time window, and token counts. Results are audit findings and never overwrite settlement facts.
      parameters:
        - $ref: "#/components/parameters/IDPath"
        - name: adapter
          in: query
          schema: { type: string }
        - name: model
          in: query
          schema: { type: string }
        - name: start_at
          in: query
          schema: { type: integer, format: int64 }
        - name: end_at
          in: query
          schema: { type: integer, format: int64 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 5000 }
        - name: status_code
          in: query
          schema: { type: integer, minimum: 100, maximum: 599 }
      responses:
        "200": { $ref: "#/components/responses/APIResponse" }
...
  /api/v1/admin/group/{id}/channels:
    get:
      tags: [Admin Groups]
      summary: List group channels
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200": { $ref: "#/components/responses/APIResponse" }
    put:
      tags: [Admin Groups]
      summary: Update group channels
      parameters:
        - $ref: "#/components/parameters/IDPath"
      requestBody: { $ref: "#/components/requestBodies/JSONBody" }
      responses:
        "200": { $ref: "#/components/responses/APIResponse" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or API token
    CookieAuth:
      type: apiKey
      in: cookie
      name: session
    TopupCallbackToken:
      type: apiKey
      in: header
      name: X-Topup-Callback-Token
  parameters:
    IDPath:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
    Keyword:
      name: keyword
      in: query
      schema: { type: string }
...
  requestBodies:
    JSONBody:
      required: false
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
...
  responses:
    APIResponse:
      description: Router standard response envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
    PaginatedAPIResponse:
      description: Router standard paginated response envelope
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/APIResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PaginatedData"
...
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
        error:
          nullable: true
      additionalProperties: true
    PaginatedData:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
      additionalProperties: true
...